Indent-line function and single tempdir
There's now a function-valued variable can be set to customize the
indentation behavior. This is convenient because you can freely use one
from another language in your config.

Also, now using a single temp file for latex compilation instead of
making a new one each time. This *might* even make the compilation
faster, if it's only changing part of the circuit. Not sure about that,
though.