Options and more precise buffer setup
Added more of the command-line options. Also, the preview buffer will
only be set up as a pdf buffer if the options are set to proceed through
code generation. This will leave it as a text buffer, even when we are
set to LaTeX mode, if we only continue to e.g. the analysis phase.